Free camera navigation (wasd+mouse) working

+++ Engine/Code/platform.h
@@ -55,8 +55,8 @@ enum ButtonState
 
 struct Input
 {
-    glm::ivec2  mousePos;
-    glm::ivec2  mouseDelta;
+    glm::vec2   mousePos;
+    glm::vec2   mouseDelta;
     ButtonState mouseButtons[MOUSE_BUTTON_COUNT];
     ButtonState keys[KEY_COUNT];
 };
